+++ android/app/src/main/kotlin/com/thepmsquare/navidrome_client/MainActivity.kt
@@ -1,5 +1,6 @@
 package com.thepmsquare.navidrome_client
 
-import io.flutter.embedding.android.FlutterActivity
+import com.ryanheise.audioservice.AudioServiceActivity
 
-class MainActivity : FlutterActivity()
+class MainActivity : AudioServiceActivity() {
+}