+++ lib/services/session_service.dart
@@ -14,6 +14,7 @@ class SessionService {
   static const String _keyAutoDownloadLruEvict = 'auto_download_lru_evict';
   static const String _keyHomeSections = 'home_sections';
   static const String _keyLastVersion = 'last_version';
+  static const String _keyLegacyHeadsetMultiClick = 'legacy_headset_multi_click';
 
   static final SessionService _instance = SessionService._internal();
   factory SessionService() => _instance;
@@ -162,6 +163,16 @@ class SessionService {
   Future<void> setAutoDownloadLruEvict(bool value) async {
     final prefs = await _getPrefs;
     await prefs.setBool(_keyAutoDownloadLruEvict, value);
+  }
+
+  Future<bool> get legacyHeadsetMultiClick async {
+    final prefs = await _getPrefs;
+    return prefs.getBool(_keyLegacyHeadsetMultiClick) ?? true;
+  }
+
+  Future<void> setLegacyHeadsetMultiClick(bool value) async {
+    final prefs = await _getPrefs;
+    await prefs.setBool(_keyLegacyHeadsetMultiClick, value);
   }
 
   // ---------------------------------------------------------------------------